Fixes for compiling with VC6.

+++ src/yajl/yajl_gen.c
@@ -142,7 +142,7 @@ yajl_gen_free(yajl_gen g)
         yajl_buf_append(g->buf, "\n", 1);        
     
 yajl_gen_status
-yajl_gen_integer(yajl_gen g, long long int number)
+yajl_gen_integer(yajl_gen g, long int number)
 {
     char i[32];
     ENSURE_VALID_STATE; ENSURE_NOT_KEY; INSERT_SEP; INSERT_WHITESPACE;
